requirements.txt
python-dotenv
google-cloud-secret-manager